requirements.txt
py-jwt
fastapi
python-gnupg
sqlalchemy
uvicorn
jinja
asyncio